+++ src/planner/constify_now.c
@@ -33,9 +33,8 @@
  * - Var >= now() - Interval
  * - Var >= now() + Interval
  *
- * Additionally Interval needs to be Const and not contain day or month
- * components as those would be affected by timezone, which can change
- * between executions of a prepared statement.
+ * Additionally Interval needs to be Const and not contain a month
+ * component.
  */
 static const Dimension *
 get_hypertable_dimension(Oid relid)
@@ -96,11 +95,11 @@ is_valid_now_expr(OpExpr *op, List *rtable)
 
 	Interval *offset = DatumGetIntervalP(c->constvalue);
 	/*
-	 * We don't consider day or month intervals safe here as
+	 * We don't consider month intervals safe here as
 	 * they are affected by timezones and therefore not
 	 * safe to evaluate during planning.
 	 */
-	if (offset->day != 0 || offset->month != 0)
+	if (offset->month != 0)
 		return false;
 
 	return true;
@@ -145,12 +144,33 @@ constify_now_expr(PlannerInfo *root, OpExpr *op)
 	else
 	{
 		OpExpr *op_inner = lsecond_node(OpExpr, op->args);
+		Const *const_offset = lsecond_node(Const, op_inner->args);
+		Assert(const_offset->consttype == INTERVALOID);
+		Interval *offset = DatumGetIntervalP(const_offset->constvalue);
 		/*
 		 * Sanity check that this is a supported expression. We should never
 		 * end here if it isn't since this is checked in is_valid_now_expr.
 		 */
 		Assert(linitial_node(FuncExpr, op_inner->args)->funcid == F_NOW);
 		linitial(op_inner->args) = make_now_const();
+
+		/*
+		 * If the interval has a day component then the calculation needs
+		 * to take into account daylight saving time switches and thereby a
+		 * day would not always be exactly 24 hours. We mitigate this by
+		 * adding a safety buffer to account for these dst switches when
+		 * dealing with intervals with day component. These calculations
+		 * will be repeated with exact values during execution.
+		 * Since dst switches seem to range between -1 and 2 hours we set
+		 * the safety buffer to 4 hours.
+		 */
+		if (offset->day != 0)
+		{
+			Const *now = linitial_node(Const, op_inner->args);
+			int64 now_value = DatumGetInt64(now->constvalue);
+			now_value -= 4 * USECS_PER_HOUR;
+			now->constvalue = Int64GetDatum(now_value);
+		}
 
 		/*
 		 * Normally estimate_expression_value is not safe to use during planning