+++ CHANGELOG.md
@@ -6,6 +6,10 @@ accidentally triggering the load of a previous DB version.**
 
 ## Unreleased
 
+**Features**
+* #4374 Remove constified now() constraints from plan
+* #4393 Support intervals with day component when constifying now()
+
 ## 2.7.0 (2022-05-24)
 
 This release adds major new features since the 2.6.1 release.